refactor(path): 경로 정보를 의미하는 Path 객체 패키지 이동 및 통과한 전체 노선 ID 필드 추가
